sdoh/requirements.txt
grpcio==1.27.2
grpcio-tools==1.27.2
protobuf==3.12.2

allennlp==0.9.0
calmsize==0.1.3
gensim-bz2-nsml==3.8.0
GPUtil==1.4.0
joblib==0.13.2
more-itertools==8.7.0
nltk==3.4.1
numpy==1.20.0
pandas==0.24.2
pytorch-memlab==0.0.2
regex==2019.4.14
scikit-learn==0.21.3
seaborn==0.10.1
torch==1.4.0
torchvision==0.5.0
tqdm==4.31.1